fix(models): unify duplicate Agent/AgentRun/ProjectMember models, unbreak app startup
Unbreak main (broken since merge 8dbc815): import app failed, no tests could run.

- models/agent.py: single canonical Agent mapping for agents table; workspace
  runtime + collaboration (owner/kind/capabilities) fields merged as nullable
  superset; AgentStatus extended with paused/offline/disabled
- models/agent_run.py: single AgentRun mapping; trigger-engine fields relaxed
  to nullable, collaboration fields (task_id/assignment_id/status/
  AgentRunStatus) added
- models/agent_core.py: keep TaskAssignment/RunLog/enums/helpers; drop dup
  Agent/AgentRun classes, re-export canonical ones
- models/audit_project.py + project_member.py: ProjectMember dedup; ADMIN
  role value + get_role/can RBAC helpers on canonical model; ProjectRole alias
- models/__init__.py: export all collaboration symbols (workflow, channels,
  protocols, sandbox, conflicts, experience_reputation, task_collab, audit)
- models/agent.py: compat re-export layer so legacy 'from models.agent import X'
  call sites keep working
- models/workflow.py: disambiguate WorkflowStep.workflow FK (workflow_id vs
  sub_workflow_id)
- channels/protocols/experience_reputation: fix Integer FKs to BigInteger
  tasks.id (MySQL FK constraint error 3780)
- core/auth + consumers: JWT identity must be string (flask-jwt-extended 4.6 /
  PyJWT 2.10 sub validation); coerce on read for legacy int tokens
- api/sse.py, api/agents/dashboard.py: future annotations for Py3.9
- conftest: pbkdf2 fallback when hashlib.scrypt unavailable (LibreSSL)
- feat(task): add tasks.dod + human_intervention_count (ACR metric) columns
  and TaskEvidenceRecord model with migration (P1.2 DoD groundwork)
- tests: align smoke tests with post-merge API surface (invite endpoints,
  workspace agents listing, 403 in allowed statuses); 94/94 passing

+++ models/channels.py
@@ -157,7 +157,7 @@ class KnowledgeEntry(BaseModel):
     domain = Column(String(100), comment="Knowledge domain (e.g. 'python', 'frontend', 'devops')")
     tags = Column(JSON, default=list, comment="List of tags for categorization")
     entry_type = Column(String(50), default="insight", comment="Type: insight, pattern, solution, reference, rule")
-    source_task_id = Column(Integer, ForeignKey("tasks.id"), nullable=True, comment="Task that generated this knowledge")
+    source_task_id = Column(BigInteger, ForeignKey("tasks.id"), nullable=True, comment="Task that generated this knowledge")
     source_type = Column(String(50), default="manual", comment="Source: manual, auto_extracted, imported, shared")
     confidence = Column(Float, default=1.0, comment="Confidence score 0.0-1.0")
     access_count = Column(Integer, default=0, comment="How many times this entry has been accessed")
